Add files via upload
This project demonstrates the concept of inheritance and the use of the base keyword in C#. The base keyword is used to call the constructor or methods of a base class from a derived class. In the example, the Car class inherits from the Vehicle class, and the base constructer is called to initialize properties of the base class.